Large refactoring and fully implementing UI In order to handle splitting, I decided a Hand class was necessary. Player and Dealer share lots of common functionality in their hands. Hand allows a unified way for player and dealer to create their hands.
I changed all the names of the cards to match the rank and suit.
Now I can use the names in the path to display images in the scene.

All text but win/loss has been removed from the terminal.

The code appears much more readible now. Comments still need to be added.